+++ changelog.md
@@ -25,6 +25,8 @@ Codion Change Log
 - EntityEditEvents removed, EntityEditModel.EditEvents added along with editEvents().
 - EntitySearchModel.Builder.condition()  now accepts a null argument.
 - DefaultEntitySearchModel, improved exception message when refreshing in case the condition supplier returns null or a condition for the incorrect type.
+### is.codion.swing.common.ui
+- DefaultFilterTableColumnModel, validation for unique model indexes and identifiers added.
 ### is.codion.swing.framework.model
 - EntityComboBoxModel.Builder.condition() now accepts a null argument.
 - DefaultEntityComboBoxModel, improved exception message when refreshing in case the condition supplier returns null or a condition for the incorrect type.